kprobes: Delay removing probe from the hash table when unregistering
When unregistering a kprobe, we disarm the probe (i.e., remove the
breakpoint) and remove the kprobe structure from the hash table without
delay. Due to this, kprobe handlers need to account for a scenario where
a breakpoint is hit, but the kprobe structure is not to be found in the
hash table. This is often done by reading the instruction at the
breakpoint address and checking if it is actually the breakpoint
instruction we installed or not.

However, this is not always simple and architectures have had to employ
workarounds when handling such scenarios. Commit 21f8b2fa3ca5b0
("powerpc/kprobes: Ignore traps that happened in real mode") and commit
9ed5df69b79a22 ("powerpc/kprobes: Use probe_address() to read
instructions") are a few recent examples of some nasty problems due to
this.

Fix this by using synchorize_rcu() between disarming the probe and
removing it from the hash table.

Signed-off-by: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>