powerpc64/bpf: Add support for bpf trampolines
Add support for bpf_arch_text_poke() and arch_prepare_bpf_trampoline()
for powerpc64 -mprofile-kernel.

We set aside space for two stubs at the beginning of each bpf program.
These stubs are used if having to branch to locations outside the range
of a branch instruction.

BPF Trampolines adhere to the powerpc64 -mprofile-kernel ABI since these
need to attach to ftrace locations using ftrace direct attach. Due to
this, bpf_arch_text_poke() patches two instructions: 'mflr r0' and 'bl'
for BPF_MOD_CALL. The trampoline code itself closely follows the x86
implementation.

This is early RFC to get feedback on the approach. This series depends
on some of the other BPF JIT fixes and enhancements posted previously,
as well as on ftrace direct enablement on powerpc which has also been
posted in the past.

Signed-off-by: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>